requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyarrow>=14.0.1
python-dotenv>=1.0.0

# Data processing
joblib>=1.3.0
pyyaml>=6.0.1
requests>=2.31.0

# Optional dependencies
ipython>=8.14.0  # For development
jupyter>=1.0.0   # For development
